app testing framework implemented (cherry picked from commit 8421db3)
