+++ tethys_sdk/testing.py
@@ -1,0 +1,11 @@
+"""
+********************************************************************************
+* Name: stores.py
+* Author: Shawn Crawley
+* Created On: 18 November 2016
+* Copyright: (c) Brigham Young University 2016
+* License: BSD 2-Clause
+********************************************************************************
+"""
+# DO NOT ERASE
+from tethys_apps.base.testing import *
